NAV → NAV_TAB. Hold for nav, tap for tab.
For the longest time, the leftmost thumb key was only used to
momentarily toggle the nav layer but considering how often I hit the tab key and how conveniently placed that thumb key is, it's a good idea to give it a dual function. It allows me to stop diagonally moving my pinky to autocomplete stuff and whatnot.