Experimenting with different MS_WH keys positions
Ideally the goal is to find a setup to replace the mouse wheel because
scrolling with a mouse wheel isn't the most ergo thing you can do.
It's a very repetitive motion that requires a relatively big amount of
motion from the finger doing the scrolling on the mouse wheel.